# src/main/resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "REST API for Big Bad Con",
    "version": "1.0",
    "title": "Big Bad Con data service REST API",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Jeremy Tidwell",
      "url": "https://www.bigbadcon.com/",
      "email": "necanthrope@gmail.com"
    },
    "license": {
      "name": "Apache License Version 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8091",
  "basePath": "/",
...
  "paths": {
...
    "/events/category/{category}": {
      "get": {
        "tags": [
          "events-controller"
        ],
        "summary": "Find all events for a category.",
        "operationId": "findByCategoryUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "description": "category",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Events"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/events/me": {
      "get": {
        "tags": [
          "events-controller"
        ],
        "summary": "Find all events for the current user.",
        "operationId": "findForMeUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/events/year/{year}/all": {
      "get": {
        "tags": [
          "events-for-year-controller"
        ],
        "summary": "Find all events for a given year.",
        "operationId": "findByYearUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "description": "year",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Events"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events/year/{year}/category/{category}": {
      "get": {
        "tags": [
          "events-for-year-controller"
        ],
        "summary": "Find all events for a given year, for a category.",
        "operationId": "findByCategoryUsingGET_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "description": "year",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "category",
            "in": "path",
            "description": "category",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Events"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events/year/{year}/count": {
      "get": {
        "tags": [
          "events-for-year-controller"
        ],
        "summary": "Count of events for a given year",
        "operationId": "getCountForYearUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "description": "year",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events/year/{year}/me": {
      "get": {
        "tags": [
          "events-for-year-controller"
        ],
        "summary": "Find all events for the current user for a given year.",
        "operationId": "findForMeForYearUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "description": "year",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events/year/{year}/page/{length}/{offset}": {
      "get": {
        "tags": [
          "events-for-year-controller"
        ],
        "summary": "Find all events for a given year.",
        "operationId": "findByYearUsingGET_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "description": "year",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "length",
            "in": "path",
            "description": "length",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "offset",
            "in": "path",
            "description": "offset",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Events"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Postmeta": {
      "type": "object",
      "properties": {
        "metaId": {
          "type": "integer",
          "format": "int64"
        },
        "metaKey": {
          "type": "string"
        },
        "metaValue": {
          "type": "string"
        },
        "postId": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
    "UsersShort": {
      "type": "object",
      "properties": {
        "bbcUserFavorites": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BbcUserFavorites"
          }
        },
        "bookingComment": {
          "type": "string"
        },
        "displayName": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "Time": {
      "type": "object",
      "properties": {
        "date": {
          "type": "integer",
          "format": "int32"
        },
        "day": {
          "type": "integer",
          "format": "int32"
        },
        "hours": {
          "type": "integer",
          "format": "int32"
        },
        "minutes": {
          "type": "integer",
          "format": "int32"
        },
        "month": {
          "type": "integer",
          "format": "int32"
        },
        "seconds": {
          "type": "integer",
          "format": "int32"
        },
        "time": {
          "type": "integer",
          "format": "int64"
        },
        "timezoneOffset": {
          "type": "integer",
          "format": "int32"
        },
        "year": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "Timestamp": {
      "type": "object",
      "properties": {
        "date": {
          "type": "integer",
          "format": "int32"
        },
        "day": {
          "type": "integer",
          "format": "int32"
        },
        "hours": {
          "type": "integer",
          "format": "int32"
        },
        "minutes": {
          "type": "integer",
          "format": "int32"
        },
        "month": {
          "type": "integer",
          "format": "int32"
        },
        "nanos": {
          "type": "integer",
          "format": "int32"
        },
        "seconds": {
          "type": "integer",
          "format": "int32"
        },
        "time": {
          "type": "integer",
          "format": "int64"
        },
        "timezoneOffset": {
          "type": "integer",
          "format": "int32"
        },
        "year": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "Events": {
      "type": "object",
      "properties": {
        "blogId": {
          "type": "integer",
          "format": "int64"
        },
        "bookingExempt": {
          "type": "integer",
          "format": "int32"
        },
        "bookings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Bookings"
          }
        },
        "categories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BbcEventCategoriesShort"
          }
        },
        "eventAllDay": {
          "type": "integer",
          "format": "int32"
        },
        "eventAttributes": {
          "type": "string"
        },
        "eventCategoryId": {
          "type": "integer",
          "format": "int64"
        },
        "eventDateCreated": {
          "$ref": "#/definitions/Timestamp"
        },
        "eventDateModified": {
          "$ref": "#/definitions/Timestamp"
        },
        "eventEndDate": {
          "type": "string",
          "format": "date"
        },
        "eventEndTime": {
          "$ref": "#/definitions/Time"
        },
        "eventId": {
          "type": "integer",
          "format": "int64"
        },
        "eventName": {
          "type": "string"
        },
        "eventOwner": {
          "$ref": "#/definitions/UsersShort"
        },
        "eventPrivate": {
          "type": "string",
          "format": "byte"
        },
        "eventRoom": {
          "type": "string"
        },
        "eventRsvp": {
          "type": "string",
          "format": "byte"
        },
        "eventRsvpDate": {
          "type": "string",
          "format": "date"
        },
        "eventRsvpSpaces": {
          "type": "integer",
          "format": "int32"
        },
        "eventRsvpTime": {
          "$ref": "#/definitions/Time"
        },
        "eventSlug": {
          "type": "string"
        },
        "eventSpaces": {
          "type": "integer",
          "format": "int32"
        },
        "eventStartDate": {
          "type": "string",
          "format": "date"
        },
        "eventStartTime": {
          "$ref": "#/definitions/Time"
        },
        "eventStatus": {
          "type": "integer",
          "format": "int32"
        },
        "eventTable": {
          "type": "string"
        },
        "groupId": {
          "type": "integer",
          "format": "int64"
        },
        "lastUpdated": {
          "$ref": "#/definitions/Timestamp"
        },
        "locationId": {
          "type": "integer",
          "format": "int64"
        },
        "metadata": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Postmeta"
          }
        },
        "postContent": {
          "type": "string"
        },
        "postId": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
    "BbcUserFavorites": {
      "type": "object",
      "properties": {
        "bbcUserFavoritesId": {
          "type": "integer",
          "format": "int64"
        },
        "createDate": {
          "$ref": "#/definitions/Timestamp"
        },
        "eventId": {
          "type": "integer",
          "format": "int64"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "updateDate": {
          "$ref": "#/definitions/Timestamp"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
    "BbcEventCategoriesShort": {
      "type": "object",
      "properties": {
        "eventId": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "postId": {
          "type": "integer",
          "format": "int64"
        },
        "slug": {
          "type": "string"
        }
      }
    },
...
    "Bookings": {
      "type": "object",
      "properties": {
        "bookingComment": {
          "type": "string"
        },
        "bookingDate": {
          "$ref": "#/definitions/Timestamp"
        },
        "bookingId": {
          "type": "integer",
          "format": "int64"
        },
        "bookingMeta": {
          "type": "string"
        },
        "bookingPrice": {
          "type": "number"
        },
        "bookingSpaces": {
          "type": "integer",
          "format": "int32"
        },
        "bookingStatus": {
          "type": "integer",
          "format": "int32"
        },
        "bookingTaxRate": {
          "type": "number"
        },
        "bookingTaxes": {
          "type": "number"
        },
        "event": {
          "$ref": "#/definitions/Events"
        },
        "lastUpdated": {
          "$ref": "#/definitions/Timestamp"
        },
        "user": {
          "$ref": "#/definitions/UsersShort"
        }
      }
    }
  }
}